[Bug report] front end flowchart
Bug:
try the file 'test_space/test_nested_if.py';
>> not all node connect to the 'End' node(and it should be, in logic)

---

Display issue?
about desion statement, 'if' and 'elif', has its own block;
but for 'else' case, it display as 'Yes' or 'No', for 'No' on the edge, not a block

So When the animation play, the animation show what block is really executed,
for 'else' statement, who doesn't has side effect like:

```python
a = 1
b = 2
c = 3

if(a == 1)://                <<--- from here
    if(b == 1):
        ...
    elif(b==2)://            <<--- to here
        if(c==1):
            print('4')
        elif(c==2):
            print('5')
        else:
            print('6')
    else:
        ...
```
in animation, it look like a 'jump'; pass a intermediate node

but i'm not sure should we add that node into animation,
or highlight the edge(true, false), or keep like now.